[update] App.jsx Main.jsx StartForm.jsx StartFormSelect.jsx follow airbnb eslint
